test: add integration tests for token storage
add integration test for token storage
