+++ packages/integration-tests/src/helpers/experience/index.ts
@@ -2,7 +2,7 @@
  * @fileoverview This file contains the successful interaction flow helper functions that use the experience APIs.
  */
 
-import { type SocialUserInfo } from '@logto/connector-kit';
+import { type TokenResponse, type SocialUserInfo } from '@logto/connector-kit';
 import {
   InteractionEvent,
   SignInIdentifier,
@@ -162,15 +162,18 @@ export const identifyUserWithEmailVerificationCode = async (
 
 /**
  *
- * @param socialUserInfo The social user info that will be returned by the social connector.
+ * @param socialUserInfo The social user info and token response that will be returned by the social connector.
  * @param registerNewUser Optional. If true, the user will be registered if the user does not exist, otherwise a error will be thrown if the user does not exist.
  */
 export const signInWithSocial = async (
   connectorId: string,
-  socialUserInfo: SocialUserInfo,
+  socialUserInfo: SocialUserInfo & {
+    tokenResponse?: TokenResponse;
+  },
   options?: {
     registerNewUser?: boolean;
     linkSocial?: boolean;
+    tokenResponse?: TokenResponse;
   }
 ) => {
   const state = 'state';
