+++ packages/integration-tests/src/tests/api/experience-api/sign-in-interaction/social.test.ts
@@ -1,39 +1,34 @@
 import { ConnectorType } from '@logto/connector-kit';
-import { InteractionEvent, SignInIdentifier } from '@logto/schemas';
 import { generateStandardId } from '@logto/shared';
 
 import {
   mockSocialConnectorId,
   mockSocialConnectorTarget,
 } from '#src/__mocks__/connectors-mock.js';
-import { deleteUser, getUser } from '#src/api/admin-user.js';
-import { initExperienceClient, logoutClient, processSession } from '#src/helpers/client.js';
+import { deleteUser, getUser, getUserIdentityTokenSetRecord } from '#src/api/admin-user.js';
+import { updateConnectorConfig } from '#src/api/connector.js';
+import { isDevFeaturesEnabled } from '#src/constants.js';
 import {
   clearConnectorsByTypes,
   setEmailConnector,
   setSocialConnector,
 } from '#src/helpers/connector.js';
 import { signInWithSocial } from '#src/helpers/experience/index.js';
-import {
-  successFullyCreateSocialVerification,
-  successFullyVerifySocialAuthorization,
-} from '#src/helpers/experience/social-verification.js';
-import {
-  successfullySendVerificationCode,
-  successfullyVerifyVerificationCode,
-} from '#src/helpers/experience/verification-code.js';
-import { expectRejects } from '#src/helpers/index.js';
-import {
-  enableAllPasswordSignInMethods,
-  enableAllVerificationCodeSignInMethods,
-} from '#src/helpers/sign-in-experience.js';
+import { enableAllPasswordSignInMethods } from '#src/helpers/sign-in-experience.js';
 import { generateNewUser } from '#src/helpers/user.js';
-import { generateEmail, generateUsername } from '#src/utils.js';
+import { generateEmail } from '#src/utils.js';
 
 describe('social sign-in and sign-up', () => {
   const connectorIdMap = new Map<string, string>();
   const socialUserId = generateStandardId();
   const email = generateEmail();
+  const mockTokenResponse = {
+    id_token: 'mid_token',
+    access_token: 'access_token',
+    token_type: 'bearer',
+    scope: 'openid ',
+    expires_in: 3600,
+  };
 
   beforeAll(async () => {
     await enableAllPasswordSignInMethods({
@@ -44,6 +39,15 @@ describe('social sign-in and sign-up', () => {
     await clearConnectorsByTypes([ConnectorType.Social, ConnectorType.Email]);
 
     const { id: socialConnectorId } = await setSocialConnector();
+
+    // Enable token storage
+    // TODO: Remove this once we have token storage enabled
+    if (isDevFeaturesEnabled) {
+      await updateConnectorConfig(socialConnectorId, {
+        enableTokenStorage: true,
+      });
+    }
+
     await setEmailConnector();
     connectorIdMap.set(mockSocialConnectorId, socialConnectorId);
   });
@@ -58,6 +62,7 @@ describe('social sign-in and sign-up', () => {
       {
         id: socialUserId,
         email,
+        tokenResponse: mockTokenResponse,
       },
       {
         registerNewUser: true,
@@ -66,6 +71,13 @@ describe('social sign-in and sign-up', () => {
 
     const { primaryEmail } = await getUser(userId);
     expect(primaryEmail).toBe(email);
+
+    // TODO: Remove this once we have token storage enabled
+    if (isDevFeaturesEnabled) {
+      const tokenSetRecord = await getUserIdentityTokenSetRecord(userId, mockSocialConnectorTarget);
+      expect(tokenSetRecord).not.toBeNull();
+      expect(tokenSetRecord.metadata.scope).toBe(mockTokenResponse.scope);
+    }
   });
 
   it('should successfully sign-up with social but not sync email if the email is registered by another user', async () => {
@@ -98,11 +110,21 @@ describe('social sign-in and sign-up', () => {
       id: socialUserId,
       email,
       name: 'John Doe',
+      tokenResponse: {
+        ...mockTokenResponse,
+        scope: 'openid profile email',
+      },
     });
 
     const { name } = await getUser(userId);
 
     expect(name).toBe('John Doe');
+
+    if (isDevFeaturesEnabled) {
+      const tokenSetRecord = await getUserIdentityTokenSetRecord(userId, mockSocialConnectorTarget);
+      expect(tokenSetRecord).not.toBeNull();
+      expect(tokenSetRecord.metadata.scope).toBe('openid profile email');
+    }
 
     await deleteUser(userId);
   });
@@ -118,6 +140,10 @@ describe('social sign-in and sign-up', () => {
         id: socialUserId,
         email: userProfile.primaryEmail,
         name: 'Foo Bar',
+        tokenResponse: {
+          ...mockTokenResponse,
+          scope: 'openid profile phone',
+        },
       },
       { linkSocial: true }
     );
@@ -127,225 +153,12 @@ describe('social sign-in and sign-up', () => {
     expect(identities[mockSocialConnectorTarget]).toBeTruthy();
     expect(name).toBe('Foo Bar');
 
+    if (isDevFeaturesEnabled) {
+      const tokenSetRecord = await getUserIdentityTokenSetRecord(userId, mockSocialConnectorTarget);
+      expect(tokenSetRecord).not.toBeNull();
+      expect(tokenSetRecord.metadata.scope).toBe('openid profile phone');
+    }
+
     await deleteUser(userId);
-  });
-
-  describe('fulfill missing user profile', () => {
-    const state = 'state';
-    const redirectUri = 'http://localhost:3000';
-
-    it('should successfully sign-up with social and fulfill missing username', async () => {
-      const connectorId = connectorIdMap.get(mockSocialConnectorId)!;
-
-      await enableAllPasswordSignInMethods({
-        identifiers: [SignInIdentifier.Username],
-        password: true,
-        verify: false,
-      });
-
-      const client = await initExperienceClient();
-      const { verificationId } = await successFullyCreateSocialVerification(client, connectorId, {
-        redirectUri,
-        state,
-      });
-
-      await successFullyVerifySocialAuthorization(client, connectorId, {
-        verificationId,
-        connectorData: {
-          state,
-          redirectUri,
-          code: 'fake_code',
-          userId: generateStandardId(),
-        },
-      });
-
-      await expectRejects(client.identifyUser({ verificationId }), {
-        code: 'user.identity_not_exist',
-        status: 404,
-      });
-
-      await client.updateInteractionEvent({ interactionEvent: InteractionEvent.Register });
-
-      await expectRejects(client.identifyUser({ verificationId }), {
-        code: 'user.missing_profile',
-        status: 422,
-      });
-
-      await client.updateProfile({ type: SignInIdentifier.Username, value: generateUsername() });
-
-      await client.identifyUser();
-
-      const { redirectTo } = await client.submitInteraction();
-      const userId = await processSession(client, redirectTo);
-      await logoutClient(client);
-      await deleteUser(userId);
-    });
-
-    describe('fulfill missing email', () => {
-      beforeAll(async () => {
-        await enableAllVerificationCodeSignInMethods({
-          identifiers: [SignInIdentifier.Email],
-          password: true,
-          verify: true,
-        });
-      });
-
-      it('should directly sync trusted email', async () => {
-        const userId = await signInWithSocial(
-          connectorIdMap.get(mockSocialConnectorId)!,
-          {
-            id: socialUserId,
-            email,
-          },
-          {
-            registerNewUser: true,
-          }
-        );
-
-        const { primaryEmail } = await getUser(userId);
-        expect(primaryEmail).toBe(email);
-
-        await deleteUser(userId);
-      });
-
-      it('should ask to provide email if no verified email is returned from social', async () => {
-        const connectorId = connectorIdMap.get(mockSocialConnectorId)!;
-
-        const client = await initExperienceClient();
-
-        const { verificationId } = await successFullyCreateSocialVerification(client, connectorId, {
-          redirectUri,
-          state,
-        });
-
-        await successFullyVerifySocialAuthorization(client, connectorId, {
-          verificationId,
-          connectorData: {
-            state,
-            redirectUri,
-            code: 'fake_code',
-            userId: generateStandardId(),
-          },
-        });
-
-        await expectRejects(client.identifyUser({ verificationId }), {
-          code: 'user.identity_not_exist',
-          status: 404,
-        });
-
-        await client.updateInteractionEvent({ interactionEvent: InteractionEvent.Register });
-
-        await expectRejects(client.identifyUser({ verificationId }), {
-          code: 'user.missing_profile',
-          status: 422,
-        });
-
-        const identifier = Object.freeze({ type: SignInIdentifier.Email, value: generateEmail() });
-
-        const { code, verificationId: emailVerificationId } =
-          await successfullySendVerificationCode(client, {
-            identifier,
-            interactionEvent: InteractionEvent.Register,
-          });
-
-        await successfullyVerifyVerificationCode(client, {
-          identifier,
-          verificationId: emailVerificationId,
-          code,
-        });
-
-        await client.updateProfile({
-          type: SignInIdentifier.Email,
-          verificationId: emailVerificationId,
-        });
-
-        await client.identifyUser();
-
-        const { redirectTo } = await client.submitInteraction();
-        const userId = await processSession(client, redirectTo);
-        await logoutClient(client);
-        await deleteUser(userId);
-      });
-
-      it('should ask to sign-in and link social if the email is already in use', async () => {
-        const { userProfile, user } = await generateNewUser({
-          primaryEmail: true,
-        });
-
-        const { primaryEmail } = userProfile;
-
-        const connectorId = connectorIdMap.get(mockSocialConnectorId)!;
-
-        const client = await initExperienceClient();
-
-        const { verificationId } = await successFullyCreateSocialVerification(client, connectorId, {
-          redirectUri,
-          state,
-        });
-
-        await successFullyVerifySocialAuthorization(client, connectorId, {
-          verificationId,
-          connectorData: {
-            state,
-            redirectUri,
-            code: 'fake_code',
-            userId: generateStandardId(),
-          },
-        });
-
-        await expectRejects(client.identifyUser({ verificationId }), {
-          code: 'user.identity_not_exist',
-          status: 404,
-        });
-
-        await client.updateInteractionEvent({ interactionEvent: InteractionEvent.Register });
-
-        await expectRejects(client.identifyUser({ verificationId }), {
-          code: 'user.missing_profile',
-          status: 422,
-        });
-
-        const identifier = Object.freeze({ type: SignInIdentifier.Email, value: primaryEmail });
-
-        const { code, verificationId: emailVerificationId } =
-          await successfullySendVerificationCode(client, {
-            identifier,
-            interactionEvent: InteractionEvent.Register,
-          });
-
-        await successfullyVerifyVerificationCode(client, {
-          identifier,
-          verificationId: emailVerificationId,
-          code,
-        });
-
-        await expectRejects(
-          client.updateProfile({
-            type: SignInIdentifier.Email,
-            verificationId: emailVerificationId,
-          }),
-          {
-            code: 'user.email_already_in_use',
-            status: 422,
-          }
-        );
-
-        await client.updateInteractionEvent({ interactionEvent: InteractionEvent.SignIn });
-
-        await client.identifyUser({ verificationId: emailVerificationId });
-        await client.updateProfile({ type: 'social', verificationId });
-
-        const { redirectTo } = await client.submitInteraction();
-        const userId = await processSession(client, redirectTo);
-        await logoutClient(client);
-
-        expect(userId).toBe(user.id);
-
-        const { identities } = await getUser(userId);
-        expect(identities[mockSocialConnectorTarget]).toBeTruthy();
-
-        await deleteUser(userId);
-      });
-    });
   });
 });
